testsuite: the instruction buffer is volatile unsigned long
main changed the builtin instruction-buffer pointer from volatile void*
to volatile unsigned long* (riscv-builtins.cc: long_unsigned_type_node
in place of void_type_node), and the rebased compiler carries that.  The
tests declared it three different ways -- volatile unsigned, volatile
unsigned int, and plain unsigned int -- none of which convert:

    sfpi_classes.h:29: error: invalid conversion from
      volatile unsigned int* to volatile long unsigned int*

136 files, now all one spelling.  On a 32-bit target unsigned long and
unsigned int are the same width, so nothing about the tests changes
except the declared type.

+++ gcc/testsuite/g++.target/riscv/tt/sfpi/autoincr-noexec-record-hoist-refuse-bh.C
@@ -16,8 +16,8 @@
 // { dg-final { scan-rtl-dump-not "Hoisted no-exec capture" "rvtt_replay" } }
 // { dg-final { scan-rtl-dump "Dst-autoincr group: bb" "rvtt_dst_autoincr" } }
 // { dg-final { scan-rtl-dump-not "mod-write-noexec-record-composition-unaudited" "rvtt_dst_autoincr" } }
-extern volatile unsigned __instrn_buffer[];
-namespace ckernel { constexpr inline volatile unsigned (&instrn_buffer)[] = ::__instrn_buffer; }
+extern volatile unsigned long __instrn_buffer[];
+namespace ckernel { constexpr inline volatile unsigned long (&instrn_buffer)[] = ::__instrn_buffer; }
 #include <sfpi.h>
 __attribute__((noinline)) void
 qk_zeta_rows_near (volatile int *sep)
