+++ gcc/testsuite/g++.target/riscv/tt/sfpi/ccrestore-loadi-velse-fire-bh.C
@@ -3,9 +3,9 @@
 // region-entry save (SFPCOMPC.md) -- audited narrowing -- so
 // candidates in BOTH arms hoist.
 // { dg-final { scan-tree-dump-times "Hoisted invariant SFPU immediate" 2 "rvtt_invariant" } }
-extern volatile unsigned __instrn_buffer[];
+extern volatile unsigned long __instrn_buffer[];
 namespace ckernel {
-constexpr inline volatile unsigned (&instrn_buffer)[] = ::__instrn_buffer;
+constexpr inline volatile unsigned long (&instrn_buffer)[] = ::__instrn_buffer;
 }
 #include <sfpi.h>
