+++ gcc/testsuite/g++.target/riscv/tt/sfpi/intnot-fire-bh.C
@@ -6,9 +6,9 @@
 // { dg-final { scan-assembler "SFPNOT\tL\[0-7\], L\[0-7\]" } }
 // { dg-final { scan-assembler-not "SFPIADD" } }
 // { dg-final { scan-assembler-not "SFPLOADI\t.*-1" } }
-extern volatile unsigned __instrn_buffer[];
+extern volatile unsigned long __instrn_buffer[];
 namespace ckernel {
-constexpr inline volatile unsigned (&instrn_buffer)[] = ::__instrn_buffer;
+constexpr inline volatile unsigned long (&instrn_buffer)[] = ::__instrn_buffer;
 }
 #include <sfpi.h>
 __attribute__((noinline)) void
