+++ gcc/testsuite/g++.target/riscv/tt/sfpi/lut-select-leafext-slot-creg-nonfloatb-keep-bh.C
@@ -15,9 +15,9 @@
 // { dg-final { scan-assembler-times "SFPMOV" 1 } }
 // { dg-final { scan-assembler "SFPMAD\tL3, L3, L3, L10" } }
 
-extern volatile unsigned __instrn_buffer[];
+extern volatile unsigned long __instrn_buffer[];
 namespace ckernel {
-constexpr inline volatile unsigned (&instrn_buffer)[] = ::__instrn_buffer;
+constexpr inline volatile unsigned long (&instrn_buffer)[] = ::__instrn_buffer;
 }
 #include <sfpi.h>
