+++ gcc/testsuite/g++.target/riscv/tt/sfpi/store-fold-x-invariant-compose-bh.C
@@ -12,8 +12,8 @@
 // immediate still hoists cleanly.
 // { dg-final { scan-tree-dump "store-fold: sank post-region store" "rvtt_store_fold" } }
 // { dg-final { scan-tree-dump "Hoisted invariant SFPU immediate" "rvtt_invariant" } }
-extern volatile unsigned __instrn_buffer[];
-namespace ckernel { constexpr inline volatile unsigned (&instrn_buffer)[] = ::__instrn_buffer; }
+extern volatile unsigned long __instrn_buffer[];
+namespace ckernel { constexpr inline volatile unsigned long (&instrn_buffer)[] = ::__instrn_buffer; }
 #include <sfpi.h>
 using namespace sfpi;
 __attribute__((noinline)) void
